pyproject.toml
[project]
name = "backend-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiokafka>=0.12.0",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.3",
    "pydantic[email]>=2.11.7",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "result>=0.17.0",
    "sentry-sdk>=2.33.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
